Adding the T junction block and RF source module.
This file extracts the coordinates supplied by a bitmap raster to define the boundary conditions and source for our FEM.

T_junction(image_path="T_junction.bmp",remove_whitespace=True)
INPUT
image.bmp -> This is the input image which I kept the same as the raster size of the simulation.
		It doesn't have to be the same size. The coordinates are extracted from the image.
		In this version there are only ideal walls and source blocks.
		In this version:
		The black pixels are the wall and are detected by finding pixels with B=0
		The 'pink' pixels are the source coordinates and are for pixels with G=127
		BMP has BGR pixel orientation.

remove_whitespace -> This subtracts the minimum value from the coordinates as an offset.
			This is an appendage for when I command the coordinates with a local handler object.

BUG Note there is a slight bug in which it only takes the minimum from the wall coordinates and applys it to
both the source and walls. If by chance a source pixel precedes a wall, then we may end up offsetting the source
off the raster. This is corrected in the C++ version.