Adding fullfigure matplotlib wrapper
Wrapper figure class to call the animations in ways "just my way". Everyone seems to have their own collection of matplotlib scripts that
are ungeneralizable. If you want it, sure come and have it. I left my treasure in on place.
Ironically this is like the fifth wrapper function that I have made for just my way. It seems to change every time. This is disposable code.